Update L1 Task7.cpp

+++ LB1/L1 Task7.cpp
@@ -5,7 +5,7 @@ int main()
 {
 	const int size = 4; // Amount of segments (We can change this value)
 
-	int segments[size]; // Create segments
+	double segments[size]; // Create segments
 	double p; // Semi-perimeter for calculations
 
 	while (true) // Loop out program
